+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/DeleteLakehouseMonitorRequest.java
@@ -1,0 +1,41 @@
+// Code generated from OpenAPI specs by Databricks SDK Generator. DO NOT EDIT.
+
+package com.databricks.sdk.service.catalog;
+
+import com.databricks.sdk.support.Generated;
+import com.databricks.sdk.support.ToStringer;
+import java.util.Objects;
+
+/** Delete a table monitor */
+@Generated
+public class DeleteLakehouseMonitorRequest {
+  /** Full name of the table. */
+  private String fullName;
+
+  public DeleteLakehouseMonitorRequest setFullName(String fullName) {
+    this.fullName = fullName;
+    return this;
+  }
+
+  public String getFullName() {
+    return fullName;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+    DeleteLakehouseMonitorRequest that = (DeleteLakehouseMonitorRequest) o;
+    return Objects.equals(fullName, that.fullName);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(fullName);
+  }
+
+  @Override
+  public String toString() {
+    return new ToStringer(DeleteLakehouseMonitorRequest.class).add("fullName", fullName).toString();
+  }
+}