+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/MonitorCronSchedulePauseStatus.java
@@ -1,0 +1,12 @@
+// Code generated from OpenAPI specs by Databricks SDK Generator. DO NOT EDIT.
+
+package com.databricks.sdk.service.catalog;
+
+import com.databricks.sdk.support.Generated;
+
+/** Whether the schedule is paused or not */
+@Generated
+public enum MonitorCronSchedulePauseStatus {
+  PAUSED,
+  UNPAUSED,
+}