+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/MonitorInfoStatus.java
@@ -1,0 +1,15 @@
+// Code generated from OpenAPI specs by Databricks SDK Generator. DO NOT EDIT.
+
+package com.databricks.sdk.service.catalog;
+
+import com.databricks.sdk.support.Generated;
+
+/** The status of the monitor. */
+@Generated
+public enum MonitorInfoStatus {
+  MONITOR_STATUS_ACTIVE,
+  MONITOR_STATUS_DELETE_PENDING,
+  MONITOR_STATUS_ERROR,
+  MONITOR_STATUS_FAILED,
+  MONITOR_STATUS_PENDING,
+}