+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/TablesAPI.java
@@ -49,6 +49,25 @@ public class TablesAPI {
     impl.delete(request);
   }
 
+  public TableExistsResponse exists(String fullName) {
+    return exists(new ExistsRequest().setFullName(fullName));
+  }
+
+  /**
+   * Get boolean reflecting if table exists.
+   *
+   * <p>Gets if a table exists in the metastore for a specific catalog and schema. The caller must
+   * satisfy one of the following requirements: * Be a metastore admin * Be the owner of the parent
+   * catalog * Be the owner of the parent schema and have the USE_CATALOG privilege on the parent
+   * catalog * Have the **USE_CATALOG** privilege on the parent catalog and the **USE_SCHEMA**
+   * privilege on the parent schema, and either be the table owner or have the SELECT privilege on
+   * the table. * Have BROWSE privilege on the parent catalog * Have BROWSE privilege on the parent
+   * schema.
+   */
+  public TableExistsResponse exists(ExistsRequest request) {
+    return impl.exists(request);
+  }
+
   public TableInfo get(String fullName) {
     return get(new GetTableRequest().setFullName(fullName));
   }
@@ -56,10 +75,11 @@ public class TablesAPI {
   /**
    * Get a table.
    *
-   * <p>Gets a table from the metastore for a specific catalog and schema. The caller must be a
-   * metastore admin, be the owner of the table and have the **USE_CATALOG** privilege on the parent
-   * catalog and the **USE_SCHEMA** privilege on the parent schema, or be the owner of the table and
-   * have the **SELECT** privilege on it as well.
+   * <p>Gets a table from the metastore for a specific catalog and schema. The caller must satisfy
+   * one of the following requirements: * Be a metastore admin * Be the owner of the parent catalog
+   * * Be the owner of the parent schema and have the USE_CATALOG privilege on the parent catalog *
+   * Have the **USE_CATALOG** privilege on the parent catalog and the **USE_SCHEMA** privilege on
+   * the parent schema, and either be the table owner or have the SELECT privilege on the table.
    */
   public TableInfo get(GetTableRequest request) {
     return impl.get(request);