+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/compute/ClustersService.java
@@ -36,7 +36,9 @@ public interface ClustersService {
   /**
    * Change cluster owner.
    *
-   * <p>Change the owner of the cluster. You must be an admin to perform this operation.
+   * <p>Change the owner of the cluster. You must be an admin and the cluster must be terminated to
+   * perform this operation. The service principal application ID can be supplied as an argument to
+   * `owner_username`.
    */
   void changeOwner(ChangeClusterOwner changeClusterOwner);