+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/compute/GlobalInitScriptsAPI.java
@@ -75,7 +75,7 @@ public class GlobalInitScriptsAPI {
    *
    * <p>Get a list of all global init scripts for this workspace. This returns all properties for
    * each script but **not** the script contents. To retrieve the contents of a script, use the [get
-   * a global init script](#operation/get-script) operation.
+   * a global init script](:method:globalinitscripts/get) operation.
    */
   public Iterable<GlobalInitScriptDetails> list() {
     return impl.list().getScripts();