+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/compute/Policy.java
@@ -41,7 +41,10 @@ public class Policy {
   @JsonProperty("is_default")
   private Boolean isDefault;
 
-  /** A list of libraries to be installed on the next cluster restart that uses this policy. */
+  /**
+   * A list of libraries to be installed on the next cluster restart that uses this policy. The
+   * maximum number of libraries is 500.
+   */
   @JsonProperty("libraries")
   private Collection<Library> libraries;