+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/workspace/RepoAccessControlRequest.java
@@ -17,10 +17,7 @@ public class RepoAccessControlRequest {
   @JsonProperty("permission_level")
   private RepoPermissionLevel permissionLevel;
 
-  /**
-   * Application ID of an active service principal. Setting this field requires the
-   * `servicePrincipal/user` role.
-   */
+  /** application ID of a service principal */
   @JsonProperty("service_principal_name")
   private String servicePrincipalName;